+++ swift/container/updater.py
@@ -47,7 +47,7 @@ class ContainerUpdater(Daemon):
         self.devices = conf.get('devices', '/srv/node')
         self.mount_check = config_true_value(conf.get('mount_check', 'true'))
         self.swift_dir = conf.get('swift_dir', '/etc/swift')
-        self.interval = int(conf.get('interval', 300))
+        self.interval = float(conf.get('interval', 300))
         self.account_ring = None
         self.concurrency = int(conf.get('concurrency', 4))
         if 'slowdown' in conf:
